:recycle: Refactored chat, keystore and jokes mocks into shared util library
