+++ backend/libs/shared/util/src/index.ts
@@ -3,4 +3,6 @@ export * from './lib/chat.mock';
 export * from './lib/global-exception.filter';
 export * from './lib/http-response.mock';
 export * from './lib/identifier.mock';
+export * from './lib/joke.mock';
+export * from './lib/keystore.mock';
 export * from './lib/logger.middleware';
