+++ backend/libs/shared/util/src/lib/joke.mock.ts
@@ -1,0 +1,6 @@
+import { JokeDto } from "@backend/shared-types";
+
+export const mockJokeDto = (mock: Partial<JokeDto>): JokeDto => ({
+  text: mock.text || 'mockText',
+  language: mock.language || 'mockLanguage',
+});
